app/static-signs/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
mediapipe==0.8.9.1
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
SpeechRecognition==3.8.1
absl-py==0.13.0 